feat: Extract Desktop Policies into a dashboard section
Open-Work-Snacks-Run: f1c67cf1-5d8d-4caf-a437-871ee54d6db7:8dd6821bc19a84f5139f241a2d0c82496e124f5de86061522f745c5f803aaa48
