+++ ee/apps/den-web/tests/cloud-super-admins-role-hierarchy.test.ts
@@ -100,7 +100,7 @@ describe("cloud super-admin role hierarchy", () => {
     const diagnostics = read("../app/(den)/dashboard/_components/diagnostics-screen.tsx");
     const diagnosticCard = read("../app/(den)/dashboard/_components/egress-diagnostics-card.tsx");
     const brand = read("../app/(den)/dashboard/_components/brand-appearance-screen.tsx");
-    const desktopPolicies = read("../app/(den)/dashboard/_components/desktop-policies-screen.tsx");
+    const desktopPolicies = read("../app/(den)/dashboard/_components/desktop-policies-section.tsx");
     const desktopPolicyEditor = read("../app/(den)/dashboard/_components/desktop-policy-editor-screen.tsx");
     const billing = read("../app/(den)/dashboard/_components/billing-dashboard-screen.tsx");
     const apiKeys = read("../app/(den)/dashboard/_components/api-keys-screen.tsx");
